Support debian package and docker image testing

+++ Jenkinsfiles/Jenkinsfile-stratum-bcm-test-master.groovy
@@ -47,19 +47,41 @@ pipeline {
                         node("${TEST_DRIVER}") {
                             def WORKSPACE = pwd()
                             stage("Start Stratum on ${SWITCH_NAME}") {
-                                sh returnStdout: false, label: "Starting Stratum with image ${DOCKER_IMAGE}:${DOCKER_IMAGE_TAG}", script: """
+                                sh returnStdout: false, label: "Copy Config Files", script: """
                                     sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "mkdir -p $CONFIG_DIR"
                                     sshpass -p $SWITCH_CREDS_PSW scp ${stratum_configs_dir}/dummy_serdes_db.pb.txt $SWITCH_CREDS_USR@$SWITCH_IP:${CONFIG_DIR}/dummy_serdes_db.pb.txt
                                     sshpass -p $SWITCH_CREDS_PSW scp ${stratum_configs_dir}/bcm_hardware_specs.pb.txt $SWITCH_CREDS_USR@$SWITCH_IP:${CONFIG_DIR}/bcm_hardware_specs.pb.txt
                                     sshpass -p $SWITCH_CREDS_PSW scp -r ${stratum_configs_dir}/${SWITCH_NAME} $SWITCH_CREDS_USR@$SWITCH_IP:${CONFIG_DIR}/${SWITCH_NAME}
-                                    sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "docker stop stratum-bcm || true"
-                                    sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "docker rm stratum-bcm || true"
-                                    sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux kill-session -t CI || true"
-                                    sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux new -d -s CI || true"
-                                    sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux send-keys -t CI.0 ENTER 'docker pull ${DOCKER_IMAGE}:${DOCKER_IMAGE_TAG}' ENTER"
-                                    sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux send-keys -t CI.0 ENTER 'DOCKER_IMAGE=${DOCKER_IMAGE} DOCKER_IMAGE_TAG=${DOCKER_IMAGE_TAG} CONFIG_DIR=${CONFIG_DIR} ./start-stratum-container.sh' ENTER"
-                                    sleep 30
                                 """
+                                if (params.DEBIAN_PACKAGE_NAME != '' && params.DEBIAN_PACKAGE_PATH != '') {
+                                    sh returnStdout: false, label: "Install Stratum Debian Package", script: """
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "pkill stratum || true"
+                                        sshpass -p $SWITCH_CREDS_PSW scp ${DEBIAN_PACKAGE_PATH}/${DEBIAN_PACKAGE_NAME} $SWITCH_CREDS_USR@$SWITCH_IP:/tmp
+                                        
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux kill-session -t CI || true"
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux new -d -s CI || true"
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux send-keys -t CI.0 ENTER 'DEBIAN_PACKAGE_NAME=${DEBIAN_PACKAGE_NAME} /tmp/install_bcm_debian_package.sh; tmux wait-for -S install' C-m\\; wait-for install"
+                                    """
+                                    sh returnStdout: false, label: "Starting Stratum with Debian Package", script: """
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux send-keys -t CI.0 ENTER 'CONFIG_DIR=${CONFIG_DIR} /tmp/start-stratum.sh' C-m"
+                                        sleep 30
+                                    """
+                                } else if (params.DOCKER_IMAGE != '' && params.DOCKER_IMAGE_TAG != '') {
+                                sh returnStdout: false, label: "Starting Stratum with image ${DOCKER_IMAGE}:${DOCKER_IMAGE_TAG}", script: """
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "pkill stratum || true"
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "docker stop stratum-bcm || true"
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "docker rm stratum-bcm || true"
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux kill-session -t CI || true"
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux new -d -s CI || true"
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux send-keys -t CI.0 ENTER 'docker pull ${DOCKER_IMAGE}:${DOCKER_IMAGE_TAG}' ENTER"
+                                        sshpass -p $SWITCH_CREDS_PSW ssh $SWITCH_CREDS_USR@$SWITCH_IP "tmux send-keys -t CI.0 ENTER 'DOCKER_IMAGE=${DOCKER_IMAGE} DOCKER_IMAGE_TAG=${DOCKER_IMAGE_TAG} CONFIG_DIR=${CONFIG_DIR} ./start-stratum-container.sh' ENTER"
+                                        sleep 30
+                                    """
+                                } else {
+                                    script {
+                                        error "Invalid Parameters"
+                                    }
+                                }
                             }
                             stage('Get Test Vectors') {
                                 step([$class: 'WsCleanup'])
@@ -104,4 +126,3 @@ pipeline {
         }
     }
 }
-
